springapp/src/com/slp/domain/Product.java
package com.slp.domain;

import java.io.Serializable;

/**
 * @author sanglp
 * @create 2018-01-03 17:15
 * @desc 产品
 **/
public class Product implements Serializable {
    private String description;
    private Double price;
    private Integer id;

    public Integer getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }

    public Double getPrice() {
        return price;
    }

    public void setPrice(Double price) {
        this.price = price;
    }

    @Override
    public String toString() {
        StringBuffer buffer = new StringBuffer();
        buffer.append("Description: " + description + ";");
        buffer.append("Price: " + price);
        return buffer.toString();
    }
}
