springapp/test/com/slp/domain/ProductTests.java
package com.slp.domain;

import junit.framework.TestCase;

/**
 * @author sanglp
 * @create 2018-01-03 17:16
 * @desc 测试
 **/
public class ProductTests extends TestCase {
    private Product product;

    protected void setUp() throws Exception {
        product = new Product();
    }

    public void testSetAndGetDescription() {
        String testDescription = "aDescription";
        assertNull(product.getDescription());
        product.setDescription(testDescription);
        assertEquals(testDescription, product.getDescription());
    }

    public void testSetAndGetPrice() {
        double testPrice = 100.00;
        assertEquals(0, 0, 0);
        product.setPrice(testPrice);
        assertEquals(testPrice, product.getPrice(), 0);
    }
}
